refactor(pr-check): common out oneOf/anyOf logic
