+++ packages/schema/src/types/anyOf.ts
@@ -1,10 +1,8 @@
 import { Schema, SchemaContextCreator } from '../schema';
-import { PartialJSONSchema } from '../jsonSchemaTypes';
 import {
-  toCombinatorJSONSchemaWithDiscriminator,
   DiscriminatorMap,
   ArraySchemaType,
-  getDiscriminatedSchema,
+  createCombinatorWithDiscriminator,
 } from '../typeCombinatorUtils';
 
 export function anyOf<T extends Array<Schema<any, any>>>(
@@ -13,116 +11,16 @@ export function anyOf<T extends Array<Schema<any, any>>>(
   discriminatorField?: string
 ): Schema<ArraySchemaType<T>> {
   if (discriminatorMap && discriminatorField) {
-    return createAnyOfWithDiscriminator(
+    return createCombinatorWithDiscriminator(
       schemas,
       discriminatorMap,
-      discriminatorField
+      discriminatorField,
+      createAnyOfWithoutDiscriminator(schemas),
+      'anyOf'
     );
   } else {
     return createAnyOfWithoutDiscriminator(schemas);
   }
-}
-
-function createAnyOfWithDiscriminator<T extends Array<Schema<any, any>>>(
-  schemas: T,
-  discriminatorMap: DiscriminatorMap<T>,
-  discriminatorField: string
-): Schema<ArraySchemaType<T>> {
-  const anyOfWithoutDiscriminator = createAnyOfWithoutDiscriminator(schemas);
-  return {
-    type: () => `OneOf<${schemas.map((schema) => schema.type()).join(' | ')}>`,
-    validateBeforeMap: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.validateBeforeMap(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.validateBeforeMap(value, ctxt);
-    },
-    validateBeforeUnmap: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.validateBeforeUnmap(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.validateBeforeUnmap(value, ctxt);
-    },
-    map: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.map(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.map(value, ctxt);
-    },
-    unmap: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.unmap(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.unmap(value, ctxt);
-    },
-    validateBeforeMapXml: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.validateBeforeMapXml(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.validateBeforeMapXml(value, ctxt);
-    },
-    mapXml: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.mapXml(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.mapXml(value, ctxt);
-    },
-    unmapXml: (value, ctxt) => {
-      const discriminatedSchema = getDiscriminatedSchema(
-        value,
-        discriminatorMap,
-        discriminatorField
-      );
-      if (discriminatedSchema) {
-        return discriminatedSchema.unmapXml(value, ctxt);
-      }
-      return anyOfWithoutDiscriminator.unmapXml(value, ctxt);
-    },
-    toJSONSchema: (context): PartialJSONSchema => {
-      if (!(discriminatorMap && discriminatorField)) {
-        return {
-          anyOf: schemas.map((schema) => schema.toJSONSchema(context)),
-        };
-      }
-      return toCombinatorJSONSchemaWithDiscriminator(
-        schemas,
-        discriminatorMap,
-        discriminatorField,
-        'anyOf',
-        context
-      );
-    },
-  };
 }
 
 function createAnyOfWithoutDiscriminator<T extends Array<Schema<any, any>>>(
